+++ lib/openmaize/log_entry.ex
@@ -1,3 +1,0 @@
-defmodule Openmaize.LogEntry do
-  defstruct user: "nil", message: "", meta: []
-end